clean up: remove old components and renaming
