+++ src/routes/(app)/maps/+page.ts
@@ -1,5 +1,6 @@
-import { redirect } from "@sveltejs/kit";
+// TODO: Remove this entire folder branch
+import { redirect } from '@sveltejs/kit';
 
 export const load = async () => {
-  throw redirect(302, '/')
+	throw redirect(302, '/');
 };
