+++ src/routes/(app)/maps/[map]/+page.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
 	import StratCard from '$lib/components/containers/StratCard.svelte';
 	import MapSkeleton from '$lib/components/skeletons/MapSkeleton.svelte';
-	import Filters from '$lib/features/stratListing/components/filter/Filters.svelte';
 
 	export let data;
 
@@ -9,7 +8,6 @@
 </script>
 
 <div class="grid gap-4 w-default mt-10">
-	<Filters map={mapName} />
 	<div
 		class="grid lg:grid-cols-4 md:grid-cols-3 sm:grid-cols-2 grid-cols-1 gap-4"
 	>
